set up tox to run docs build with checker enabled
Use an environment variable to control when the extension is enabled
in conf.py.

Add settings to setup.cfg to control the build.

Add tox environment 'docs' to build HTML docs and then run spell
check.

Update the word list with more words that are used and should not
trigger a build failure.

Change-Id: I19b7e2134eb7961f9817fadf30de57ed82b855ef
Signed-off-by: Doug Hellmann <doug@doughellmann.com>