Don't fail by default
There's value in not always failing on a spelling - for example, if you
want to use the tool merely as a guideline rather than a strict check.
We could add a spelling-specific option to make failures optional, but
we already have a similar option available - Sphinx's 'warning-is-error'
option. Stop setting the builder status and instead emit a warning on
build failures.

This is breaking change so we probably want to bump the MAJOR version.

Signed-off-by: Stephen Finucane <stephen@that.guru>