Move tests out of the install to the top-level of the project
Tests don't need to be installed to users' production environments.
Generally speaking, the tests are for project developers and
contributors, not for library users. As such, these files go unused
after install. Not including these files in the install is a bit
friendlier to library users as it reduces wheel size and amount of
installed code to production environments.

The tests are now at the top-level of the project so they can continue
to be used by project contributors to test changes and features.

The tests are still included in the sdist, so downstream packagers can
still run the tests to verify their package.

With the test files outside of the package directory, the "omit"
configuration in coveragerc is now unnecessary and so has been removed.