Simplify and improve tox configuration
Remove `usedevelop = True`. It is better to test the installed package
than an in-place one. Testing against the installed version is more
robust, representative, and catches some packaging issues.

Replace `deps = .[test]` with tox's builtin `extras` configuration
option. For details on this features, see:
https://tox.readthedocs.io/en/latest/config.html#conf-extras.

Use `skip_install = true` for the linter environment. Installing
sphinxcontrib-spelling to a virtual environment is unnecessary for
flake8 purposes. This saves a small amount of time when running tox as
there are fewer operations to perform. For details on this features,
see: https://tox.readthedocs.io/en/latest/config.html#conf-skip_install.
As the package installation no longer occurs, using the extras won't
apply so the flake8 dependency has moved to the tox.ini file.

Remove `deps = .[docs]` as the "docs" extras does not exist. As the
package itself requires Sphinx, there is no need for another extras.