Switch CI from Travis to GitHub actions
GitHub actions produces faster results than Travis, resulting in faster
feedback for contributors.

The "basepython" tox configuration option was dropped as GitHub actions
default to using the latest Python version which is now 3.9. This avoid
additional churn and configuration in both actions and tox when a new
Python is released.

The PyPI publisher was not ported as I do not have a way to test it. But
should be possible to port using:

https://github.com/marketplace/actions/pypi-publish

Fixes #107