filters: fix python filename handling
If the word looks like a python module filename, strip the extension
to avoid the side-effect of actually importing the module. This
prevents, for example, 'setup.py' triggering an import of the setup
module during a doc build, which makes it look like Sphinx is
complaining about a commandline argument.

Fixes #142

Signed-off-by: Doug Hellmann <doug@doughellmann.com>