tox: pass PYENCHANT_LIBRARY_PATH variable through to commands
On macOS it can be a little tricky to set up libenchant if your
default python does not match the one used by Homebrew for the
library. Setting the variable to point to the library fixes that, but
we don't want to set it in this file for everyone so use 'passenv' to
tell tox to pass the setting through when running the commands for
each env.