Fix spelling:word role rendering verbatim
A line like:
    This :spelling:word`foobared` role needs to be fixed.

would be rendered in the output exactly like that, instead of what it
should be:
    This foobared role needs to be fixed.

Turns out the cause is quite simple: the nodes.Text object from docutils
does not hold both the raw text and the "real" text, so it is being
passed the raw text, which includes the role name as we see it.

Add a test case to cover this as well.

Fixes #188